6. Size is not grandeur, and territory does not make a nation

+++ 0x00-hello_world/6-size.c
@@ -11,10 +11,10 @@ int main(void)
 	long long int c;
 	char d;
 	float f;
-	printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(d));
-	printf("Size of an int: %lu byte(s)\n", (unsigned long)sizeof(a));
-	printf("size of a long int: %lu byte(s)\n", (unsigned long)sizeof(b));
-	printf("size of a long long int %lu byte(s)\n", (unsigned long)sizeof(c));
-	printf("size of a float: %lu byte(s)\n", (unsigned long)sizeof(f));
+	printf("Size of a char: %lu byte(s)\n", sizeof(d));
+	printf("Size of an int: %lu byte(s)\n", sizeof(a));
+	printf("size of a long int: %lu byte(s)\n", sizeof(b));
+	printf("size of a long long int %lu byte(s)\n", sizeof(c));
+	printf("size of a float: %lu byte(s)\n", sizeof(f));
 	return (0);
 }
